FIX: stripe apis (#390)
* fixed stripe_create_payment_intent api
added stripe_create_customer_id, check_payment, stripe_payment_history_list apis
fixed create_trade and FulfillTrade functions
fixed execute_recipe (insufficient coin balance issue)
added field extrainfo in get_recipe
changed output issues in create_recipe (to fix afti's issue with Conoutput denom)

* Fixed checkPayment api response params

* Delete PYLONSCLI_DOC.md

* Delete REST_SERVER_DOC.md

* Delete TESTNET3N.md

* Delete init_accounts.local

* Delete init_accounts.local(run_for_pylon).sh

* Delete init_accounts.local.sh

* Delete init_accounts.sh

* fix some minor issues based on Alex's comments

* fix lint issues, unit testing issues

* fixed stripe_payment history

* Fixed unit testing issues

+++ x/pylons/client/rest/send_coins.go
@@ -45,7 +45,7 @@ func coinsSendHandler(cliCtx client.Context) http.HandlerFunc {
 			return
 		}
 
-		msg := types.NewMsgSendCoins(types.NewPylon(int64(req.Amount)), senderAddr.String(), receiverAddr.String())
+		msg := types.NewMsgSendCoins(types.NewPylon(req.Amount), senderAddr.String(), receiverAddr.String())
 		err = msg.ValidateBasic()
 		if err != nil {
 			rest.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
